refactor(study set): rename package
