+++ app/src/main/java/com/pwhs/quickmem/presentation/app/study_set/studies/component/StudyFlipCard.kt
@@ -1,4 +1,4 @@
-package com.pwhs.quickmem.presentation.app.study_set.study.component
+package com.pwhs.quickmem.presentation.app.study_set.studies.component
 
 import androidx.compose.animation.animateContentSize
 import androidx.compose.animation.core.FastOutSlowInEasing
