+++ app/src/main/java/com/pwhs/quickmem/presentation/app/study_set/studies/flip/FlipFlashCardUiState.kt
@@ -1,10 +1,10 @@
-package com.pwhs.quickmem.presentation.app.study_set.study.quiz
+package com.pwhs.quickmem.presentation.app.study_set.studies.flip
 
 import com.pwhs.quickmem.domain.model.color.ColorModel
 import com.pwhs.quickmem.domain.model.flashcard.FlashCardResponseModel
 import com.pwhs.quickmem.domain.model.subject.SubjectModel
 
-data class LearnFlashCardUiState(
+data class FlipFlashCardUiState(
     val isLoading: Boolean = false,
     val studySetId: String = "",
     val studySetTitle: String = "",
@@ -16,13 +16,9 @@ data class LearnFlashCardUiState(
     val currentCardIndex: Int = 0,
     val learningTime: Long = 0,
     val startTime: Long = 0,
-    val flashCardLearnRound: List<FlashCardResponseModel> = emptyList(),
-    val flashCardLearnRoundIndex: Int = 0,
-    val randomAnswers: List<RandomAnswer> = emptyList(),
-)
-
-data class RandomAnswer(
-    val answer: String = "",
-    val isCorrect: Boolean = false,
-    val imageURL: String = "",
+    val isSwipingLeft: Boolean = false,
+    val isEndOfList: Boolean = false,
+    val isSwipingRight: Boolean = false,
+    val countKnown: Int = 0,
+    val countStillLearning: Int = 0,
 )
