+++ app/src/main/java/com/pwhs/quickmem/presentation/app/study_set/studies/flip/FlipFlashCardViewModel.kt
@@ -1,4 +1,4 @@
-package com.pwhs.quickmem.presentation.app.study_set.study.flip_flashcard
+package com.pwhs.quickmem.presentation.app.study_set.studies.flip
 
 import android.app.Application
 import android.media.MediaPlayer
