+++ app/src/main/java/com/pwhs/quickmem/presentation/app/study_set/studies/quiz/LearnByQuizUiEvent.kt
@@ -1,0 +1,4 @@
+package com.pwhs.quickmem.presentation.app.study_set.studies.quiz
+
+sealed class LearnByQuizUiEvent {
+}
