+++ app/src/main/java/com/pwhs/quickmem/presentation/app/study_set/studies/quiz/LearnByQuizUiState.kt
@@ -1,10 +1,10 @@
-package com.pwhs.quickmem.presentation.app.study_set.study.flip_flashcard
+package com.pwhs.quickmem.presentation.app.study_set.studies.quiz
 
 import com.pwhs.quickmem.domain.model.color.ColorModel
 import com.pwhs.quickmem.domain.model.flashcard.FlashCardResponseModel
 import com.pwhs.quickmem.domain.model.subject.SubjectModel
 
-data class FlipFlashCardUiState(
+data class LearnFlashCardUiState(
     val isLoading: Boolean = false,
     val studySetId: String = "",
     val studySetTitle: String = "",
@@ -16,9 +16,13 @@ data class FlipFlashCardUiState(
     val currentCardIndex: Int = 0,
     val learningTime: Long = 0,
     val startTime: Long = 0,
-    val isSwipingLeft: Boolean = false,
-    val isEndOfList: Boolean = false,
-    val isSwipingRight: Boolean = false,
-    val countKnown: Int = 0,
-    val countStillLearning: Int = 0,
+    val flashCardLearnRound: List<FlashCardResponseModel> = emptyList(),
+    val flashCardLearnRoundIndex: Int = 0,
+    val randomAnswers: List<RandomAnswer> = emptyList(),
+)
+
+data class RandomAnswer(
+    val answer: String = "",
+    val isCorrect: Boolean = false,
+    val imageURL: String = "",
 )
