+++ app/src/main/java/com/pwhs/quickmem/presentation/app/study_set/studies/true_false/LearnByTrueFalseScreen.kt
@@ -1,4 +1,4 @@
-package com.pwhs.quickmem.presentation.app.study_set.study.truefalse
+package com.pwhs.quickmem.presentation.app.study_set.studies.true_false
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
