--- app/src/main/java/com/pwhs/quickmem/presentation/app/study_set/study/quiz/LearnByQuizUiEvent.kt
@@ -1,4 +1,0 @@
-package com.pwhs.quickmem.presentation.app.study_set.study.quiz
-
-sealed class LearnByQuizUiEvent {
-}
